Move ticket CRUD operations to a separate file
